+++ sakura_core/charset/codechecker.h
@@ -410,7 +410,7 @@ inline int GuessEucjpCharsz( const char uc_ ){
 /* --- ローカル文字コードチェック */
 int CheckSjisChar( const char*, const int, ECharSet* );
 int CheckEucjpChar( const char*, const int, ECharSet* );
-int DetectJisEscseq( const char*, const int, EMyJisEscseq* ); // JIS エスケープシーケンス検出器
+int DetectJisEscseq( const char*, const size_t, EMyJisEscseq* ); // JIS エスケープシーケンス検出器
 int _CheckJisAnyPart( const char*, const int, const char **ppNextChar, EMyJisEscseq *peNextEsc, int *pnErrorCount, const int nType );
 enum EJisChecker{
 	JISCHECK_ASCII7,
